Add configurable proxy server and restrict security group egress
This CR is to tighten up security in the VPC to align with the recommendations
we are making for customers.
Specifically, all internet access should be required to go through
an internets proxy, security groups should provide least privilege
required by instances, and VPC endpoints should be used to access
AWS service endpoints without touching the internet.

Do not want to refer to the proxy server by IP address so
added Route53 private dns for resolving the proxy server's ip address.

Add vpc endpoints

Add vpc flow logs to help debug network issues.

Add load balancer security group. It was using Scheduler SG but
should really have it's own.

Add proxy vars to configuration so that they can be used by the application
to launch new instances and set the proxy correctly.

Updated configuration scripts so that they can easily be rerun if they fail.
This was necessary while debugging the proxy server and the security
groups.

Added KeyName to compute nodes so can be easily sshed to by admins.

cr https://code.amazon.com/reviews/CR-34688131

+++ source/scripts/Scheduler.sh
@@ -1,18 +1,23 @@
 #!/bin/bash -xe
 
-source /etc/environment
-source /root/config.cfg
-
 if [ $# -lt 2 ]
   then
+    echo "usage: $0 EFS_DATA EFS_APPS"
     exit 1
 fi
 
-# Install SSM
-yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
-systemctl enable amazon-ssm-agent
-systemctl restart amazon-ssm-agent
+set -ex
 
+source /etc/environment
+source /root/config.cfg
+source /etc/profile.d/proxy.sh
+
+# Install SSM
+if ! yum list amazon-ssm-agent; then
+    yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+fi
+systemctl enable amazon-ssm-agent || true
+systemctl restart amazon-ssm-agent
 
 mkdir -p /apps/soca/$SOCA_CONFIGURATION
 EFS_DATA=$1
@@ -35,8 +40,8 @@ yum install -y $(echo ${OPENLDAP_SERVER_PKGS[*]})
 yum install -y $(echo ${SSSD_PKGS[*]})
 
 # Mount EFS
-mkdir /apps
-mkdir /data
+mkdir -p /apps
+mkdir -p /data
 echo "$EFS_DATA:/ /data/ nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
 echo "$EFS_APPS:/ /apps nfs4 nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2,noresvport 0 0" >> /etc/fstab
 mount -a
@@ -137,71 +142,78 @@ EOF
 systemctl enable pbs
 systemctl start pbs
 
-# Default Server config
-/opt/pbs/bin/qmgr -c "create node $SERVER_HOSTNAME_ALT"
-/opt/pbs/bin/qmgr -c "set node $SERVER_HOSTNAME_ALT queue = workq"
-/opt/pbs/bin/qmgr -c "set server flatuid=true"
-/opt/pbs/bin/qmgr -c "set server job_history_enable=1"
-/opt/pbs/bin/qmgr -c "set server job_history_duration = 00:01:00"
-/opt/pbs/bin/qmgr -c "set server scheduler_iteration = 30"
-/opt/pbs/bin/qmgr -c "set server max_concurrent_provision = 5000"
+if [ ! -e /var/lib/cloud/instance/sem/pbs_server_config_done ]; then
+    # Default Server config
+    /opt/pbs/bin/qmgr -c "create node $SERVER_HOSTNAME_ALT"
+    /opt/pbs/bin/qmgr -c "set node $SERVER_HOSTNAME_ALT queue = workq"
+    /opt/pbs/bin/qmgr -c "set server flatuid=true"
+    /opt/pbs/bin/qmgr -c "set server job_history_enable=1"
+    /opt/pbs/bin/qmgr -c "set server job_history_duration = 00:01:00"
+    /opt/pbs/bin/qmgr -c "set server scheduler_iteration = 30"
+    /opt/pbs/bin/qmgr -c "set server max_concurrent_provision = 5000"
+    touch /var/lib/cloud/instance/sem/pbs_server_config_done
+fi
 
-# Default Queue Config
-/opt/pbs/bin/qmgr -c "create queue low"
-/opt/pbs/bin/qmgr -c "set queue low queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue low started = True"
-/opt/pbs/bin/qmgr -c "set queue low enabled = True"
-/opt/pbs/bin/qmgr -c "set queue low default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue normal"
-/opt/pbs/bin/qmgr -c "set queue normal queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue normal started = True"
-/opt/pbs/bin/qmgr -c "set queue normal enabled = True"
-/opt/pbs/bin/qmgr -c "set queue normal default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue high"
-/opt/pbs/bin/qmgr -c "set queue high queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue high started = True"
-/opt/pbs/bin/qmgr -c "set queue high enabled = True"
-/opt/pbs/bin/qmgr -c "set queue high default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue desktop"
-/opt/pbs/bin/qmgr -c "set queue desktop queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue desktop started = True"
-/opt/pbs/bin/qmgr -c "set queue desktop enabled = True"
-/opt/pbs/bin/qmgr -c "set queue desktop default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue test"
-/opt/pbs/bin/qmgr -c "set queue test queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue test started = True"
-/opt/pbs/bin/qmgr -c "set queue test enabled = True"
-/opt/pbs/bin/qmgr -c "set queue test default_chunk.compute_node=tbd"
-/opt/pbs/bin/qmgr -c "create queue alwayson"
-/opt/pbs/bin/qmgr -c "set queue alwayson queue_type = Execution"
-/opt/pbs/bin/qmgr -c "set queue alwayson started = True"
-/opt/pbs/bin/qmgr -c "set queue alwayson enabled = True"
-/opt/pbs/bin/qmgr -c  "set server default_queue = normal"
+if [ ! -e /var/lib/cloud/instance/sem/pbs_queue_config_done ]; then
+    # Default Queue Config
+    /opt/pbs/bin/qmgr -c "create queue low"
+    /opt/pbs/bin/qmgr -c "set queue low queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue low started = True"
+    /opt/pbs/bin/qmgr -c "set queue low enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue low default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue normal"
+    /opt/pbs/bin/qmgr -c "set queue normal queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue normal started = True"
+    /opt/pbs/bin/qmgr -c "set queue normal enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue normal default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue high"
+    /opt/pbs/bin/qmgr -c "set queue high queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue high started = True"
+    /opt/pbs/bin/qmgr -c "set queue high enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue high default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue desktop"
+    /opt/pbs/bin/qmgr -c "set queue desktop queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue desktop started = True"
+    /opt/pbs/bin/qmgr -c "set queue desktop enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue desktop default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue test"
+    /opt/pbs/bin/qmgr -c "set queue test queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue test started = True"
+    /opt/pbs/bin/qmgr -c "set queue test enabled = True"
+    /opt/pbs/bin/qmgr -c "set queue test default_chunk.compute_node=tbd"
+    /opt/pbs/bin/qmgr -c "create queue alwayson"
+    /opt/pbs/bin/qmgr -c "set queue alwayson queue_type = Execution"
+    /opt/pbs/bin/qmgr -c "set queue alwayson started = True"
+    /opt/pbs/bin/qmgr -c "set queue alwayson enabled = True"
+    /opt/pbs/bin/qmgr -c  "set server default_queue = normal"
+    touch /var/lib/cloud/instance/sem/pbs_queue_config_done
+fi
 
 # Add compute_node to list of required resource
 sed -i 's/resources: "ncpus, mem, arch, host, vnode, aoe, eoe"/resources: "ncpus, mem, arch, host, vnode, aoe, eoe, compute_node"/g' /var/spool/pbs/sched_priv/sched_config
 
-# Configure Ldap
-systemctl enable slapd
-systemctl start slapd
+if [ ! -e /var/lib/cloud/instance/sem/ldap_configured ]; then
+    # Configure Ldap
+    systemctl enable slapd
+    systemctl start slapd
 
-MASTER_LDAP_PASSWORD=$(slappasswd -g)
-MASTER_LDAP_PASSWORD_ENCRYPTED=$(/sbin/slappasswd -s $MASTER_LDAP_PASSWORD -h "{SSHA}")
-echo -n "admin" > /root/OpenLdapAdminUsername.txt
-echo -n $MASTER_LDAP_PASSWORD > /root/OpenLdapAdminPassword.txt
-chmod 600 /root/OpenLdapAdminPassword.txt
-echo "URI ldap://$SERVER_HOSTNAME" >> /etc/openldap/ldap.conf
-echo "BASE $LDAP_BASE" >> /etc/openldap/ldap.conf
+    MASTER_LDAP_PASSWORD=$(slappasswd -g)
+    MASTER_LDAP_PASSWORD_ENCRYPTED=$(/sbin/slappasswd -s $MASTER_LDAP_PASSWORD -h "{SSHA}")
+    echo -n "admin" > /root/OpenLdapAdminUsername.txt
+    echo -n $MASTER_LDAP_PASSWORD > /root/OpenLdapAdminPassword.txt
+    chmod 600 /root/OpenLdapAdminPassword.txt
+    echo "URI ldap://$SERVER_HOSTNAME" >> /etc/openldap/ldap.conf
+    echo "BASE $LDAP_BASE" >> /etc/openldap/ldap.conf
 
-# Generate 10y certificate for ldaps
-openssl req -new -newkey rsa:2048 -days 3650 -nodes -x509 \
-    -subj "/C=US/ST=California/L=Sunnyvale/O=Aligo/CN=$SERVER_HOSTNAME" \
-    -keyout /etc/openldap/certs/soca.key -out /etc/openldap/certs/soca.crt
+    # Generate 10y certificate for ldaps
+    openssl req -new -newkey rsa:2048 -days 3650 -nodes -x509 \
+        -subj "/C=US/ST=California/L=Sunnyvale/O=Aligo/CN=$SERVER_HOSTNAME" \
+        -keyout /etc/openldap/certs/soca.key -out /etc/openldap/certs/soca.crt
 
-chown ldap:ldap /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
-chmod 600 /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
+    chown ldap:ldap /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
+    chmod 600 /etc/openldap/certs/soca.key /etc/openldap/certs/soca.crt
 
-echo -e "
+    echo -e "
 dn: olcDatabase={2}hdb,cn=config
 changetype: modify
 replace: olcSuffix
@@ -218,7 +230,7 @@ replace: olcRootPW
 olcRootPW: $MASTER_LDAP_PASSWORD_ENCRYPTED
 " > db.ldif
 
-echo -e "
+    echo -e "
 dn: cn=config
 changetype: modify
 replace: olcTLSCertificateFile
@@ -228,7 +240,7 @@ replace: olcTLSCertificateKeyFile
 olcTLSCertificateKeyFile: /etc/openldap/certs/soca.key
 " > update_ssl_cert.ldif
 
-echo -e "
+    echo -e "
 dn: olcDatabase={2}hdb,cn=config
 changetype: modify
 replace: olcAccess
@@ -238,7 +250,7 @@ add: olcAccess
 olcAccess: {1}to * by dn.base="gidNumber=0+uidNumber=0,cn=peercred,cn=external,cn=auth" write by dn.base="ou=admins,$LDAP_BASE" write by * read
 " > change_user_password.ldif
 
-echo -e "
+    echo -e "
 dn: cn=sudo,cn=schema,cn=config
 objectClass: olcSchemaConfig
 cn: sudo
@@ -252,15 +264,15 @@ olcAttributeTypes: ( 1.3.6.1.4.1.15953.9.1.7 NAME 'sudoRunAsGroup' DESC 'Group(s
 olcObjectClasses: ( 1.3.6.1.4.1.15953.9.2.1 NAME 'sudoRole' SUP top STRUCTURAL DESC 'Sudoer Entries' MUST ( cn ) MAY ( sudoUser $ sudoHost $ sudoCommand $ sudoRunAs $ sudoRunAsUser $ sudoRunAsGroup $ sudoOption $ description ) )
 " > sudoers.ldif
 
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f db.ldif
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f update_ssl_cert.ldif
-/bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f change_user_password.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f sudoers.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/cosine.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/nis.ldif
-/bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/inetorgperson.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f db.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f update_ssl_cert.ldif
+    /bin/ldapmodify -Y EXTERNAL -H ldapi:/// -f change_user_password.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f sudoers.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/cosine.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/nis.ldif
+    /bin/ldapadd -Y EXTERNAL -H ldapi:/// -f /etc/openldap/schema/inetorgperson.ldif
 
-echo -e "
+    echo -e "
 dn: $LDAP_BASE
 dc: soca
 objectClass: top
@@ -287,24 +299,24 @@ objectClass: organizationalUnit
 ou: Group
 " > base.ldif
 
-/bin/ldapadd -x -W -y /root/OpenLdapAdminPassword.txt -D "cn=admin,$LDAP_BASE" -f base.ldif
+    /bin/ldapadd -x -W -y /root/OpenLdapAdminPassword.txt -D "cn=admin,$LDAP_BASE" -f base.ldif
 
-authconfig \
-    --enablesssd \
-    --enablesssdauth \
-    --enableldap \
-    --enableldapauth \
-    --ldapserver="ldap://$SERVER_HOSTNAME" \
-    --ldapbasedn="$LDAP_BASE" \
-    --enablelocauthorize \
-    --enablemkhomedir \
-    --enablecachecreds \
-    --updateall
+    authconfig \
+        --enablesssd \
+        --enablesssdauth \
+        --enableldap \
+        --enableldapauth \
+        --ldapserver="ldap://$SERVER_HOSTNAME" \
+        --ldapbasedn="$LDAP_BASE" \
+        --enablelocauthorize \
+        --enablemkhomedir \
+        --enablecachecreds \
+        --updateall
 
-echo "sudoers: files sss" >> /etc/nsswitch.conf
+    echo "sudoers: files sss" >> /etc/nsswitch.conf
 
-# Configure SSSD
-echo -e "[domain/default]
+    # Configure SSSD
+    echo -e "[domain/default]
 enumerate = True
 autofs_provider = ldap
 cache_credentials = True
@@ -343,10 +355,13 @@ ldap_sudo_smart_refresh_interval=3600
 [ifp]
 
 [secrets]" > /etc/sssd/sssd.conf
-chmod 600 /etc/sssd/sssd.conf
+    chmod 600 /etc/sssd/sssd.conf
 
-systemctl enable sssd
-systemctl restart sssd
+    systemctl enable sssd
+    systemctl restart sssd
+
+    touch /var/lib/cloud/instance/sem/ldap_configured
+fi
 
 # Disable SELINUX
 sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' /etc/selinux/config
@@ -361,13 +376,15 @@ echo "UserKnownHostsFile /dev/null" >> /etc/ssh/ssh_config
 
 # PBS hooks still run on python2 env, so some packages are required
 EASY_INSTALL=$(which easy_install-2.7)
-$EASY_INSTALL yaml python-ldap boto3
+$EASY_INSTALL yaml || true
+$EASY_INSTALL python-ldap boto3
 
-# Configure NTP
-yum remove -y ntp
-yum install -y chrony
-mv /etc/chrony.conf  /etc/chrony.conf.original
-echo -e """
+if [ ! -e /var/lib/cloud/instance/sem/ntp_configured ]; then
+    # Configure NTP
+    yum remove -y ntp
+    yum install -y chrony
+    mv /etc/chrony.conf  /etc/chrony.conf.original
+    echo -e """
 # use the local instance NTP service, if available
 server 169.254.169.123 prefer iburst minpoll 4 maxpoll 4
 
@@ -394,7 +411,9 @@ logdir /var/log/chrony
 dumponexit
 dumpdir /var/run/chrony
 """ > /etc/chrony.conf
-systemctl enable chronyd
+    systemctl enable chronyd
+    touch /var/lib/cloud/instance/sem/ntp_configured
+fi
 
 # Disable ulimit
 echo -e  "
@@ -404,4 +423,11 @@ echo -e  "
 
 # Reboot to ensure SELINUX is disabled
 # Note: Upon reboot, SchedulerPostReboot.sh script will be executed and will finalize scheduler configuration
-reboot
+if [ ! -e /var/lib/cloud/instance/sem/rebooted ] || ! needs-restarting -r; then
+    touch /var/lib/cloud/instance/sem/rebooted
+    reboot
+else
+    aws s3 cp s3://${SOCA_INSTALL_BUCKET}/${SOCA_INSTALL_BUCKET_FOLDER}/scripts/SchedulerPostReboot.sh /root/
+    chmod +x /root/SchedulerPostReboot.sh
+    /root/call-SchedulerPostReboot.sh >> /root/PostRebootConfig.log 2>&1
+fi
