+++ source/scripts/SchedulerPostReboot.sh
@@ -1,10 +1,17 @@
 #!/bin/bash -xe
 
+if [ $# -lt 4 ]
+  then
+    echo "usage: $0 S3Bucket S3InstallFolder UserName UserPassword"
+    exit 1
+fi
+
 source /etc/environment
 source /root/config.cfg
+source /etc/profile.d/proxy.sh
 
 # First flush the current crontab to prevent this script to run on the next reboot
-crontab -r
+crontab -r || true
 
 # Copy  Aligo scripts file structure
 AWS=$(which aws)
@@ -17,7 +24,8 @@ mkdir -p /apps/soca/$SOCA_CONFIGURATION/cluster_manager/logs
 chmod +x /apps/soca/$SOCA_CONFIGURATION/cluster_manager/aligoqstat.py
 
 # Generate default queue_mapping file based on default AMI choosen by customer
-cat <<EOT >> /apps/soca/$SOCA_CONFIGURATION/cluster_manager/settings/queue_mapping.yml
+if [ ! -e /apps/soca/$SOCA_CONFIGURATION/cluster_manager/settings/queue_mapping.yml ]; then
+  cat <<EOT >> /apps/soca/$SOCA_CONFIGURATION/cluster_manager/settings/queue_mapping.yml
 # This manage automatic provisioning for your queues
 # These are default values. Users can override them at job submission
 # https://awslabs.github.io/scale-out-computing-on-aws/tutorials/create-your-own-queue/
@@ -94,16 +102,18 @@ queue_type:
     #placement_group: "false"
     # .. Refer to the doc for more supported parameters
 EOT
+fi
 
 # Generate 10 years internal SSL certificate for Soca Web UI
 cd /apps/soca/$SOCA_CONFIGURATION/cluster_web_ui
-openssl req -new -newkey rsa:4096 -days 3650 -nodes -x509 \
-    -subj "/C=US/ST=California/L=Sunnyvale/CN=internal.soca.webui.cert" \
-    -keyout cert.key -out cert.crt
+if ! [ -e cert.crt ]; then
+    openssl req -new -newkey rsa:4096 -days 3650 -nodes -x509 \
+        -subj "/C=US/ST=California/L=Sunnyvale/CN=internal.soca.webui.cert" \
+        -keyout cert.key -out cert.crt
+fi
 
 # Wait for PBS to restart
 sleep 60
-
 
 ## Update PBS Hooks with the current script location
 sed -i "s/%SOCA_CONFIGURATION/$SOCA_CONFIGURATION/g" /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_acls.py
@@ -115,18 +125,21 @@ sed -i "s/%SOCA_CONFIGURATION/$SOCA_CONFIGURATION/g" /apps/soca/$SOCA_CONFIGURAT
 sed -i "s/%SOCA_CONFIGURATION/$SOCA_CONFIGURATION/g" /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/job_notifications.py
 
 # Create Default PBS hooks
-qmgr -c "create hook check_queue_acls event=queuejob"
-qmgr -c "import hook check_queue_acls application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_acls.py"
-qmgr -c "create hook check_queue_instance_types event=queuejob"
-qmgr -c "import hook check_queue_instance_types application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_instance_types.py"
-qmgr -c "create hook check_queue_restricted_parameters event=queuejob"
-qmgr -c "import hook check_queue_restricted_parameters application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_restricted_parameters.py"
-qmgr -c "create hook check_licenses_mapping event=queuejob"
-qmgr -c "import hook check_licenses_mapping application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_licenses_mapping.py"
+if [ ! -e /var/lib/cloud/instance/sem/pbs_hooks_created ]; then
+  qmgr -c "create hook check_queue_acls event=queuejob"
+  qmgr -c "import hook check_queue_acls application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_acls.py"
+  qmgr -c "create hook check_queue_instance_types event=queuejob"
+  qmgr -c "import hook check_queue_instance_types application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_instance_types.py"
+  qmgr -c "create hook check_queue_restricted_parameters event=queuejob"
+  qmgr -c "import hook check_queue_restricted_parameters application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_queue_restricted_parameters.py"
+  qmgr -c "create hook check_licenses_mapping event=queuejob"
+  qmgr -c "import hook check_licenses_mapping application/x-python default /apps/soca/$SOCA_CONFIGURATION/cluster_hooks/queuejob/check_licenses_mapping.py"
 
+  # Reload config
+  systemctl restart pbs
 
-# Reload config
-systemctl restart pbs
+  touch /var/lib/cloud/instance/sem/pbs_hooks_created
+fi
 
 # Create crontabs
 echo "
@@ -155,27 +168,32 @@ echo "
 
 # Start Web UI
 chmod +x /apps/soca/$SOCA_CONFIGURATION/cluster_web_ui/socawebui.sh
+/apps/soca/$SOCA_CONFIGURATION/cluster_web_ui/socawebui.sh stop
 /apps/soca/$SOCA_CONFIGURATION/cluster_web_ui/socawebui.sh start
 
-# When using custom AMI, the scheduler is fully operational even before SecretManager is ready. LDAP_Manager has a dependency on SecretManager so we have to wait a little bit (or create the user manually once secretmanager is available)
-MAX_ATTEMPT=5
-CURRENT_ATTEMPT=0
-# Create default LDAP user
+if [ ! -e /var/lib/cloud/instance/sem/user_created ]; then
+    # When using custom AMI, the scheduler is fully operational even before SecretManager is ready. LDAP_Manager has a dependency on SecretManager so we have to wait a little bit (or create the user manually once secretmanager is available)
+    MAX_ATTEMPT=5
+    CURRENT_ATTEMPT=0
+    # Create default LDAP user
 
-sanitized_username="$3"
-sanitized_password="$4"
+    sanitized_username="$3"
+    sanitized_password="$4"
 
-until `/apps/soca/$SOCA_CONFIGURATION/python/latest/bin/python3 /apps/soca/$SOCA_CONFIGURATION/cluster_manager/ldap_manager.py add-user -u $sanitized_username -p $sanitized_password --admin >> /dev/null 2>&1`
-do
-  echo "Unable to add new LDAP user as command failed (secret manager not ready?) Waiting 3mn ..."
-  if [[ $CURRENT_ATTEMPT -ge $MAX_ATTEMPT ]];
-  then
-    echo "Unable to create LDAP user after 5 attempts, try to run the command manually: /apps/soca/$SOCA_CONFIGURATION/python/latest/bin/python3 /apps/soca/cluster_manager/ldap_manager.py add-user -u '$3' -p '$4' --admin"
-    break
-  fi
-  sleep 180
-  ((CURRENT_ATTEMPT=CURRENT_ATTEMPT+1))
-done
+    until `/apps/soca/$SOCA_CONFIGURATION/python/latest/bin/python3 /apps/soca/$SOCA_CONFIGURATION/cluster_manager/ldap_manager.py add-user -u $sanitized_username -p $sanitized_password --admin >> /dev/null 2>&1`
+    do
+      echo "Unable to add new LDAP user as command failed (secret manager not ready?) Waiting 3mn ..."
+      if [[ $CURRENT_ATTEMPT -ge $MAX_ATTEMPT ]];
+      then
+        echo "Unable to create LDAP user after 5 attempts, try to run the command manually: /apps/soca/$SOCA_CONFIGURATION/python/latest/bin/python3 /apps/soca/cluster_manager/ldap_manager.py add-user -u '$3' -p '$4' --admin"
+        break
+      fi
+      sleep 180
+      ((CURRENT_ATTEMPT=CURRENT_ATTEMPT+1))
+    done
+
+    touch /var/lib/cloud/instance/sem/user_created
+fi
 
 # Re-enable access
 if [ "$SOCA_BASE_OS" == "amazonlinux2" ] || [ "$SOCA_BASE_OS" == "rhel7" ];
@@ -251,19 +269,23 @@ rm -rf /root/pbspro-18.1.4*
 rm -rf /root/*.sh
 rm -rf /root/config.cfg
 
-# Install OpenMPI
-# This will take a while and is not system blocking, so adding at the end of the install process
-mkdir -p /apps/soca/$SOCA_CONFIGURATION/openmpi/installer
-cd /apps/soca/$SOCA_CONFIGURATION/openmpi/installer
+if [ ! -e /var/lib/cloud/instance/sem/open_mpi_installed ]; then
+  # Install OpenMPI
+  # This will take a while and is not system blocking, so adding at the end of the install process
+  mkdir -p /apps/soca/$SOCA_CONFIGURATION/openmpi/installer
+  cd /apps/soca/$SOCA_CONFIGURATION/openmpi/installer
 
-wget $OPENMPI_URL
-if [[ $(md5sum $OPENMPI_TGZ | awk '{print $1}') != $OPENMPI_HASH ]];  then
-    echo -e "FATAL ERROR: Checksum for OpenMPI failed. File may be compromised." > /etc/motd
-    exit 1
+  wget $OPENMPI_URL
+  if [[ $(md5sum $OPENMPI_TGZ | awk '{print $1}') != $OPENMPI_HASH ]];  then
+      echo -e "FATAL ERROR: Checksum for OpenMPI failed. File may be compromised." > /etc/motd
+      exit 1
+  fi
+
+  tar xvf $OPENMPI_TGZ
+  cd openmpi-$OPENMPI_VERSION
+  ./configure --prefix=/apps/soca/$SOCA_CONFIGURATION/openmpi/$OPENMPI_VERSION
+  make
+  make install
+
+  touch /var/lib/cloud/instance/sem/open_mpi_installed
 fi
-
-tar xvf $OPENMPI_TGZ
-cd openmpi-$OPENMPI_VERSION
-./configure --prefix=/apps/soca/$SOCA_CONFIGURATION/openmpi/$OPENMPI_VERSION
-make
-make install
